backend/requirements.txt
flask
neo4j
redis
pyjwt
werkzeug
python-dotenv
flask-cors